+++ src/CsvPayload.php
@@ -3,7 +3,9 @@
 namespace Middlewares;
 
 use Interop\Http\ServerMiddleware\MiddlewareInterface;
+use InvalidArgumentException;
 use Psr\Http\Message\StreamInterface;
+use SplTempFileObject;
 
 class CsvPayload extends Payload implements MiddlewareInterface
 {
@@ -13,23 +15,97 @@ class CsvPayload extends Payload implements MiddlewareInterface
     protected $mimetype = 'text/csv';
 
     /**
+     * The field delimiter (one character only).
+     *
+     * @var string
+     */
+    protected $delimiter = ",";
+
+    /**
+     * The field enclosure (one character only).
+     *
+     * @var string
+     */
+    protected $enclosure = "\"";
+
+    /**
+     * The field escape (one character only).
+     *
+     * @var string
+     */
+    protected $escape = "\\";
+
+    /**
+     * Set Csv Control delimiter character
+     *
+     * @param string $enclosure
+     *
+     * @return self
+     */
+    public function delimiter($delimiter)
+    {
+        $this->delimiter = $this->filterControl($delimiter, 'delimiter');
+
+        return $this;
+    }
+
+    /**
+     * Set Csv Control enclosure character
+     *
+     * @param string $enclosure
+     *
+     * @return self
+     */
+    public function enclosure($enclosure)
+    {
+        $this->enclosure = $this->filterControl($enclosure, 'enclosure');
+
+        return $this;
+    }
+
+    /**
+     * Set Csv Control escape character
+     *
+     * @param string $enclosure
+     *
+     * @return self
+     */
+    public function escape($escape)
+    {
+        $this->escape = $this->filterControl($escape, 'escape');
+
+        return $this;
+    }
+
+    /**
+     * Filter Csv control character
+     *
+     * @param string $char Csv control character
+     * @param string $type Csv control character type
+     *
+     * @throws InvalidArgumentException If the Csv control character is not one character only.
+     *
+     * @return string
+     */
+    protected function filterControl($char, $type)
+    {
+        if (1 == strlen($char)) {
+            return $char;
+        }
+
+        throw new InvalidArgumentException(sprintf('The %s character must be a single character', $type));
+    }
+
+    /**
      * {@inheritdoc}
      */
     protected function parse(StreamInterface $stream)
     {
-        if ($stream->isSeekable()) {
-            $stream->rewind();
-        }
+        $csv = new SplTempFileObject();
+        $csv->fwrite((string) $stream);
+        $csv->setFlags(SplTempFileObject::READ_CSV);
+        $csv->setCsvControl($this->delimiter, $this->enclosure, $this->escape);
 
-        $resource = $stream->detach();
-        $data = [];
-
-        while (($row = fgetcsv($resource)) !== false) {
-            $data[] = $row;
-        }
-
-        fclose($resource);
-
-        return $data;
+        return iterator_to_array($csv);
     }
 }