+++ CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/) 
+The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## Next
+
+* Improve CsvPayload
+
+    - New option `delimiter()` to configure the CSV delimiter character
+    - New option `enclosure()` to configure the CSV enclosure character
+    - New option `escape()` to configure the CSV escape character
+    - `StreamInterface` fixed left undetached
 
 ## 0.3.0 - 2016-12-26